+++ src/IndoqaFela.tsx
@@ -1,6 +1,6 @@
 import {IRenderer} from 'fela'
 import React from 'react'
-import {Provider as FelaProvider} from 'react-fela'
+import {RendererProvider} from 'react-fela'
 
 export type Props = {
   renderer: IRenderer,
@@ -11,9 +11,9 @@ export class IndoqaFela extends React.Component<Props> {
 
   public render() {
     return (
-      <FelaProvider renderer={this.props.renderer}>
+      <RendererProvider renderer={this.props.renderer}>
         {this.props.children}
-      </FelaProvider>
+      </RendererProvider>
     )
   }
 }